api/pyproject.toml
[tool.poetry]
name = "kaed"
version = "0.1.0"
description = "Project description"
authors = ["kencx <ken@cheo.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
black = "^21.9b0"
mypy = "^0.910"
isort = "^5.9.3"
flake8 = "^4.0.0"
coverage = {version = "^6.0.1", extras = ["toml"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ["project"]

[tool.coverage.report]
show_missing = true

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
